feat(ThemeService): 특정 테마에 대한 예약이 존재하는 테마 삭제 시 예외 기능 구현
- 이에 따른 테스트 코드 추가

+++ src/main/java/roomescape/service/ThemeService.java
@@ -2,8 +2,10 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.Theme;
+import roomescape.exception.ExistReservationInThemeException;
 import roomescape.exception.NotExistThemeException;
 import roomescape.service.dto.input.ThemeInput;
 import roomescape.service.dto.output.ThemeOutput;
@@ -11,10 +13,12 @@ import roomescape.service.dto.output.ThemeOutput;
 @Service
 public class ThemeService {
 
-    ThemeDao themeDao;
+    final ThemeDao themeDao;
+    final ReservationDao reservationDao;
 
-    public ThemeService(ThemeDao themeDao) {
+    public ThemeService(ThemeDao themeDao, ReservationDao reservationDao) {
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
     }
 
     public ThemeOutput createTheme(ThemeInput input) {
@@ -30,6 +34,9 @@ public class ThemeService {
     public void deleteTheme(long id) {
         themeDao.find(id)
                 .orElseThrow(() -> new NotExistThemeException(String.format("%d는 없는 id 입니다.", id)));
+        if (reservationDao.isExistByThemeId(id)) {
+            throw new ExistReservationInThemeException(String.format("%d에 해당하는 예약이 있습니다.", id));
+        }
 
         themeDao.delete(id);
     }
